Catch InvalidInvariantException in middleware
Caught exceptions are now sent to a method on the command which handles them. This allows each message to handle the set of messages that it expects and return a different failure for each one.

Also changed back to returning only a single failure but a failure subclass `MultipleFailureReasons` can hold multiple failures.